fix(ci): preserve ecs service tags
We tag our services with App and Env tags set on the service for cost
allocation. It is especially useful in the AWS Cost Explorer console.
The terraform definition of our services uses the propagate_tags setting
when deploying the service, but our github actions do not by default do
the same.

This change adds the appropriate config to our github action to
propagate tags the same way our terraform code does.

Note that our cost alloctaions may be incorrectly categorizing charges
in each of these repositories from our original merge of the gh action
to now (March 28, 2025).