fix(ci): fail deployment action when deployment is rolled back
Problem: When ECS's circuit breaker mechanism rolls back a failing
deploy, the deploy task github action will exit with a success code when
infact the deployment was not successful.

In order to identify unsuccessful service updates, we want the github
action to fail. At time of writing, there is no option elicit this
behavior from the aws-actions/amazon-ecs-deploy-task-definition@v2
action. Instead, we must check that the services current task definition
is the one we attempted to roll out.

See discussion in aws-actions/amazon-ecs-deploy-task-definition#191